requirements.txt
flask
flask_limiter
flask-wtf
flask-login
flask-talisman
requests
apscheduler
PyJWT
gunicorn